+++ src/arrays/impls/mod.rs
@@ -1,0 +1,6 @@
+/// display array content
+pub mod display;
+/// iterable over array
+pub mod iter;
+/// math operators implementation
+pub mod ops;
