+++ src/arrays/macros/mod.rs
@@ -1,0 +1,2 @@
+/// ArrayTrait - Array Create functions - macro
+pub mod create;
