+++ src/arrays/manipulate.rs
@@ -1,0 +1,74 @@
+use crate::arrays::Array;
+use crate::traits::{
+    create::ArrayCreate,
+    manipulate::ArrayManipulate,
+    meta::ArrayMeta,
+    types::Numeric,
+};
+
+impl <N: Numeric> ArrayManipulate<N> for Array<N> {
+
+    fn reshape(&self, shape: Vec<usize>) -> Self {
+        assert_eq!(self.elements.len(), shape.iter().product(), "Shape must match values length");
+        Array::new(self.elements.clone(), shape)
+    }
+
+    fn ravel(&self) -> Self {
+        Array::new(self.elements.clone(), vec![self.len()])
+    }
+
+    fn for_each<F: FnMut(&N)>(&self, f: F) {
+        self.elements.iter()
+            .for_each(f)
+    }
+
+    fn for_each_e<F: FnMut(usize, &N)>(&self, mut f: F) {
+        self.elements.iter().enumerate()
+            .for_each(|(idx, item)| f(idx, item))
+    }
+
+    fn map<F: FnMut(&N) -> N>(&self, f: F) -> Self {
+        self.elements.iter()
+            .map(f)
+            .collect()
+    }
+
+    fn map_e<F: FnMut(usize, &N) -> N>(&self, mut f: F) -> Self {
+        self.elements.iter().enumerate()
+            .map(|(idx, item)| f(idx, item))
+            .collect()
+    }
+
+    fn filter<F: FnMut(&N) -> bool>(&self, mut f: F) -> Self {
+        self.elements.clone().into_iter()
+            .filter(|item| f(item))
+            .collect::<Array<N>>()
+            .ravel()
+    }
+
+    fn filter_e<F: FnMut(usize, &N) -> bool>(&self, mut f: F) -> Self {
+        self.elements.clone().into_iter().enumerate()
+            .filter(|(idx, item)| f(*idx, item))
+            .map(|i| i.1)
+            .collect::<Array<N>>()
+            .ravel()
+    }
+
+    fn filter_map<F: FnMut(&N) -> Option<N>>(&self, f: F) -> Self {
+        self.elements.clone().iter()
+            .filter_map(f)
+            .collect::<Array<N>>()
+            .ravel()
+    }
+
+    fn filter_map_e<F: FnMut(usize, &N) -> Option<N>>(&self, mut f: F) -> Self {
+        self.elements.clone().iter().enumerate()
+            .filter_map(|(idx, item)| f(idx, item))
+            .collect::<Array<N>>()
+            .ravel()
+    }
+
+    fn fold<F: FnMut(&N, &N) -> N>(&self, init: N, mut f: F) -> N {
+        self.elements.iter().fold(init, |a, b| f(&a, b))
+    }
+}
