--- src/base/mod.rs
@@ -1,17 +1,0 @@
-/// Base Array module
-pub mod base_array;
-
-/// Base Array numeric type module
-pub mod base_type;
-
-/// Compare implementations
-pub mod compare;
-
-/// Display implementation
-pub mod display;
-
-/// Ops implementation
-pub mod ops;
-
-/// Iterable implementation
-pub mod iter;
