+++ src/traits/types.rs
@@ -1,16 +1,29 @@
+use std::ops::{
+    Add, AddAssign,
+    Div, DivAssign,
+    Mul, MulAssign,
+    RangeInclusive,
+    Rem, RemAssign,
+    Sub, SubAssign,
+};
+
 use rand::Rng;
 use rand::distributions::Uniform;
 
 /// Base numeric type for array
-pub trait Numeric: Copy + Clone + PartialEq + PartialOrd + std::fmt::Display +
-std::ops::Add<Self, Output=Self> + std::ops::Sub<Self, Output=Self> +
-std::ops::Mul<Self, Output=Self> + std::ops::Div<Self, Output=Self> {
+pub trait Numeric:
+Copy + Clone + PartialEq + PartialOrd + std::fmt::Display +
+Add<Self, Output=Self> + AddAssign<Self> +
+Sub<Self, Output=Self> + SubAssign<Self> +
+Mul<Self, Output=Self> + MulAssign<Self> +
+Div<Self, Output=Self> + DivAssign<Self> +
+Rem<Self, Output=Self> + RemAssign<Self> {
     /// Zero constant value
     const ZERO: Self;
     /// One constant value
     const ONE: Self;
     /// Generate random value
-    fn rand(range: std::ops::RangeInclusive<Self>) -> Self;
+    fn rand(range: RangeInclusive<Self>) -> Self;
 }
 
 macro_rules! impl_zero_one_numeric {
@@ -19,7 +32,7 @@ macro_rules! impl_zero_one_numeric {
             const ZERO: Self = 0 as $t;
             const ONE: Self = 1 as $t;
 
-            fn rand(range: std::ops::RangeInclusive<Self>) -> $t {
+            fn rand(range: RangeInclusive<Self>) -> $t {
                 let mut rng = rand::thread_rng();
                 let value = rng.sample(&Uniform::from(range));
                 value as $t
