+++ tests/arr_rs_tests/macros/mod.rs
@@ -1,0 +1,1 @@
+mod create_test;
