+++ tests/arr_rs_tests/mod.rs
@@ -1,6 +1,8 @@
-mod array_test;
-mod compare_test;
-mod display_test;
-mod iter_test;
-mod macros_test;
-mod ops_test;
+mod create_test;
+mod indexing_test;
+mod manipulate_test;
+mod math_test;
+mod meta_test;
+
+mod impls;
+mod macros;
